+++ prefs_browser/lib/prefs_light.dart
@@ -1,0 +1,3 @@
+export 'package:tekartik_prefs/prefs_light.dart';
+export 'src/prefs_all.dart'
+    show getPrefsLightBrowser, getPrefsLightBrowserOrNull;